Global Exception Handler. Java Version updated
