+++ src/main/java/com/university/crud/springbootcrudrest/controller/StudentController.java
@@ -9,10 +9,14 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
+import org.springframework.validation.FieldError;
+import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 @RestController
 @RequestMapping("/student")
@@ -54,12 +58,5 @@ public class StudentController {
         return service.findByNameOrAge(studentName, studentAge);
     }
 
-    @ExceptionHandler(value = {Exception.class})
-    public ResponseEntity<String> getExceptions(Throwable t) {
-        System.out.println(t.getMessage());
-        System.out.println("Error Occureded");
-        t.printStackTrace();
-        return new ResponseEntity<String>("Error occured: " + t.getMessage(), HttpStatus.INTERNAL_SERVER_ERROR);
-    }
 
 }
