+++ src/main/java/com/university/crud/springbootcrudrest/model/Student.java
@@ -11,6 +11,7 @@ import javax.persistence.Id;
 import javax.validation.constraints.Email;
 import javax.validation.constraints.Min;
 import javax.validation.constraints.NotNull;
+import javax.validation.constraints.PastOrPresent;
 import java.time.LocalDate;
 
 
@@ -29,7 +30,7 @@ public class Student extends BaseEntity {
     @NotNull
     private String name;
 
-    //@FutureOrPresent
+    @PastOrPresent
     private LocalDate dob;
 
     @Email
